// asset-management/src/generated/models/Cr240_assetrequestsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr240_assetrequestscr240_status = {
  805530000: 'Pending',
  805530001: 'Approved',
  805530002: 'Rejected'
} as const;
export type Cr240_assetrequestscr240_status = keyof typeof Cr240_assetrequestscr240_status;
export const Cr240_assetrequestsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr240_assetrequestsstatecode = keyof typeof Cr240_assetrequestsstatecode;
export const Cr240_assetrequestsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr240_assetrequestsstatuscode = keyof typeof Cr240_assetrequestsstatuscode;

export interface Cr240_assetrequestsBase {
  "cr240_AssetCategoryID@odata.bind"?: string;
  cr240_assetrequestid: string;
  cr240_requestdate?: string;
  cr240_requestname?: string;
  cr240_status?: Cr240_assetrequestscr240_status;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cr240_assetrequestsstatecode;
  statuscode?: Cr240_assetrequestsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr240_assetrequests extends Cr240_assetrequestsBase {
  cr240_assetcategoryidname?: string;
  cr240_statusname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr240_assetcategoryid?: object;
  _cr240_assetcategoryid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}